backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.23.3
python-dotenv==1.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
passlib==1.7.4
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
pydantic==2.7.1
email-validator==1.3.1
